+++ src/reformat.c
@@ -423,6 +423,8 @@ avifResult avifImageRGBToYUV(avifImage * image, const avifRGBImage * rgb)
 
             avifReformatAlpha(&params);
         } else {
+            // libyuv does not import alpha from RGB to YUV so fill it
+            // independently of the value of convertedWithLibYUV.
             avifFillAlpha(&params);
         }
     }